+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- If the router notices a Peer is dead, the connection is now forcibly terminated.
+
 ## [0.3.1] - 2024-01-23
 
 ### Added