+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - If the router notices a Peer is dead, the connection is now forcibly terminated.
 - Example Systemd file.
+- Support for windows tunnels. While this works, there are no windows
+  packages yet, so this is still a very "developer experience".
 
 ### Changed